Update Student-report.cpp
The project on student report is about collecting the details of the students, for example, their name, class, roll number, attendance, and marks in each subject in a record. Whenever any parent or student wants to have a look at their grades, they can input their name and roll number, and their result will be displayed along with the other details. The result will be displayed in grades based on their marks.